requirements.txt
# CirbuildSTG Dependencies
# ========================

# Spec2RTL backend (installed as package — uses its own LLM config internally)
git+https://github.com/CirbuildProject/Cirbuild-Spec2RTL.git@main

# Agent LLM (dedicated channel, separate from spec2rtl backend)
litellm>=1.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Templating (for agent prompts)
jinja2>=3.0.0

# CLI display
rich>=13.0.0
